fix errors

+++ src/pixel_simulation/cell_model.rs
@@ -11,7 +11,7 @@ pub enum CellBehaviour {
 
 pub enum Chance {
     Always,
-    Sometimes(f32),
+    Sometimes(f64),
 }
 
 pub struct Reaction {
